+++ app/src/main/assets/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Next version
 
 * Added option to disable divider lines for tabular layout
+* Added view original signature
 
 ### 1.1832 - 2022-02-05